+++ src/App.jsx
@@ -2,15 +2,12 @@
 /* eslint-disable no-tabs */
 import React from 'react';
 import MainDisplay from './components/MainDisplay/MainDisplay';
-import useStyles from './App.style';
 
 import './App.css';
 
 function App() {
-  const styles = useStyles();
-
     return (
-        <div className={styles.root}>
+        <div>
             <MainDisplay />
         </div>
     );
