+++ src/App.style.js
@@ -1,0 +1,5 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+export default makeStyles((theme) => ({
+
+}));
