--- src/App.style.jsx
@@ -1,14 +1,0 @@
-import { makeStyles } from '@material-ui/core/styles';
-
-export default makeStyles(() => ({
-    root: {
-        '& > *': {
-            height: '100vh',
-            display: 'flex',
-            justifyContent: 'center',
-            alignItems: 'center',
-            background: '#383234',
-        },
-    },
-
-}));
