+++ src/components/MainDisplay/MainDisplay.style.js
@@ -1,0 +1,14 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+export default makeStyles(() => ({
+    image: {
+        height: '500px',
+    },
+    root: {
+        background: '#383234',
+    },
+    fullScreen: {
+        height: '100vh',
+        background: '#383234',
+    }
+}));
