--- src/components/MainDisplay/MainDisplay.style.jsx
@@ -1,13 +1,0 @@
-import { makeStyles } from '@material-ui/core/styles';
-
-
-
-export default makeStyles((theme) => ({
-    image: {
-        height: '500px',
-    },
-    button: {
-        backgroundColour: theme.palette.common.white,
-    },
-
-}));
